# api/swagger/server/restapi/resource/swagger.yml
consumes:
  - application/json
info:
  description: Thyra HTTP server API.
  title: massastation
  version: 0.2.0
produces:
  - application/json
schemes:
  - http
  - https
swagger: "2.0"
paths:
...
  /plugin-manager:
    post:
      description: Install a plugin from a given URL.
      operationId: pluginManagerInstall
      parameters:
        - in: query
          name: source
          type: string
          format: url
          required: true
          x-nullable: false
          description: URL from which to retrieve the plug-in.
      produces:
        - application/json
      responses:
        "204":
          description: Plugin successfully installed
        "400":
          description: Bad request.
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Not found.
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: Unprocessable Entity - The syntax is correct, but the server was unable to process the contained instructions.
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Internal Server Error - The server has encountered a situation it does not know how to handle.
          schema:
            $ref: "#/definitions/Error"
    get:
      description: List all installed plugins.
      operationId: pluginManagerList
      responses:
        "200":
          description: List all installed plugins.
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  description: Plugin identifier.
                  type: string
                  x-nullable: false
                name:
                  description: Plugin name.
                  type: string
                  x-nullable: false
                description:
                  description: Plugin description.
                  type: string
                  x-nullable: false
                home:
                  description: Plugin home Url.
                  type: string
                  x-nullable: false
                status:
                  description: Plugin status.
                  enum: ["Starting", "Up", "Down", "ShuttingDown", "Crashed"]
                  x-nullable: false
                logo:
                  description: Plugin logo Url.
                  type: string
                  x-nullable: false
                version:
                  description: Plugin version.
                  type: string
                  x-nullable: false

        "400":
          description: Bad request.
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Not found.
          schema:
            $ref: "#/definitions/Error"
        "422":
          description: Unprocessable Entity - The syntax is correct, but the server was unable to process the contained instructions.
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Internal Server Error - The server has encountered a situation it does not know how to handle.
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Error:
    type: object
    description: Error object.
    required:
      - code
      - message
    properties:
      code:
        description: error code.
        type: string
        x-nullable: false
      message:
        description: error message.
        type: string
        x-nullable: false
...